update fix some logic error on attendance

+++ database/migrations/2023_12_29_040722_add_column_to_events_table.php
@@ -15,7 +15,7 @@ class AddColumnToEventsTable extends Migration
     {
         Schema::table('events', function (Blueprint $table) {
             //
-            $table->unsignedBigInteger('department_id')->nullable()
+            $table->integer('department_id')->nullable()
                 ->default(0);
         });
     }
